requirements.txt
python-telegram-bot==21.1
python-dotenv==1.0.0
gspread==5.12.0
oauth2client==4.1.3
